refactor: use index file
